+++ src/Krystal/Templating/CHANGELOG.md
@@ -4,4 +4,5 @@ CHANGELOG
 1.3
 ---
 
+ * Added `StringTempalte` class with template() method to handle variables in strings
  * First version